Add assignment symbol scanning + update test descriptions
The scanner is starting to look pretty good; might have to merge some
of the find___.go files soon. I'm looking to build a regular language
but currently focusing on getting a context free grammer (CFG) scanner
working first. The CFG scanner or the strimmer could be wrapped in a
`Regulator` just like the strimmer wraps the scanner? This would allow
the separation of scanning concerns from that of grammer enforcment
ones.